feat: Add support for tracking task progress in pControl
The code changes in this commit modify the `pControl` module to include the ability to track the progress of concurrent tasks. The `start` method now accepts a callback function `concurrentDone` which is called when the concurrent tasks are finished. The callback receives two arguments: `res`, an array of objects containing the index and result of each task, and `doneSize`, the number of tasks that have finished.

This feature allows users to monitor the progress of their tasks and calculate a progress bar based on the `res` or `doneSize` values.